+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/webimport/WebImportSelectorScreen.kt
@@ -22,6 +22,7 @@ import io.github.composegears.valkyrie.jewel.tooling.PreviewTheme
 import io.github.composegears.valkyrie.sdk.compose.icons.ValkyrieIcons
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.BootstrapLogo
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.BoxLogo
+import io.github.composegears.valkyrie.sdk.compose.icons.colored.EvaLogo
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.FontAwesomeLogo
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.GoogleMaterialLogo
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.LucideLogo
@@ -29,6 +30,7 @@ import io.github.composegears.valkyrie.sdk.compose.icons.colored.RemixLogo
 import io.github.composegears.valkyrie.sdk.compose.icons.colored.TablerLogo
 import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.Bootstrap
 import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.BoxIcons
+import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.Eva
 import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.FontAwesome
 import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.GoogleMaterialSymbols
 import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.Lucide
@@ -37,6 +39,7 @@ import io.github.composegears.valkyrie.ui.screen.webimport.IconProviders.Tabler
 import io.github.composegears.valkyrie.ui.screen.webimport.material.MaterialSymbolsImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.bootstrap.BootstrapImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.boxicons.BoxIconsImportScreen
+import io.github.composegears.valkyrie.ui.screen.webimport.standard.eva.EvaImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.fontawesome.FontAwesomeImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.lucide.LucideImportScreen
 import io.github.composegears.valkyrie.ui.screen.webimport.standard.remix.RemixImportScreen
@@ -60,6 +63,7 @@ val WebImportSelectorScreen by navDestination {
                 BoxIcons -> BoxIconsImportScreen
                 FontAwesome -> FontAwesomeImportScreen
                 Tabler -> TablerImportScreen
+                Eva -> EvaImportScreen
             }
 
             navController.navigate(dest = screen)
@@ -135,6 +139,13 @@ private fun WebImportSelectorScreenUI(
                     title = stringResource("web.import.selector.tabler.title"),
                     description = stringResource("web.import.selector.tabler.description"),
                 )
+                InfoCard(
+                    onClick = { onClick(Eva) },
+                    icon = ValkyrieIcons.Colored.EvaLogo,
+                    title = stringResource("web.import.selector.eva.title"),
+                    tint = Color.Unspecified,
+                    description = stringResource("web.import.selector.eva.description"),
+                )
                 InfoText(
                     text = stringResource("web.import.selector.coming.soon"),
                     maxLines = 2,
@@ -152,6 +163,7 @@ enum class IconProviders {
     BoxIcons,
     FontAwesome,
     Tabler,
+    Eva,
 }
 
 @Preview